requirements.txt
pandas==1.2.2
requests==2.25.1
beautifulsoup4==4.9.3
joblib==1.0.1
tqdm==4.59.0
scikit-learn==0.24.1
xgboost==1.1.0
hyperopt==0.2.4
tune-sklearn==0.2.1
matplotlib==3.2.0
lightgbm==2.3.1
shap==0.35.0
alibi==0.5.6
PyALE==1.1.0
catboost==0.23.1